// client/sipejam/store/types/system.types.ts
export enum systemActionTypes {
  GET_SYSTEMS_START = "GET_SYSTEMS_START",
  GET_SYSTEMS_FAILURE = "GET_SYSTEMS_FAILURE",
  GET_SYSTEMS_SUCCESS = "GET_SYSTEMS_SUCCESS",
  GET_SYSTEM_START = "GET_SYSTEM_START",
  GET_SYSTEM_FAILURE = "GET_SYSTEM_FAILURE",
  GET_SYSTEM_SUCCESS = "GET_SYSTEM_SUCCESS",
  UPDATE_SYSTEM_START = "UPDATE_SYSTEM_START",
  UPDATE_SYSTEM_FAILURE = "UPDATE_SYSTEM_FAILURE",
  UPDATE_SYSTEM_SUCCESS = "UPDATE_SYSTEM_SUCCESS",
}